mcp_server_outlook/pyproject.toml
[project]
name = "mcp_server_outlook"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.6.0",
    "python-dotenv>=1.0.0",
    "beautifulsoup4>=4.12.0",
    "pytz>=2024.1",
    "fastmcp>=0.4.1",
]
